Add Admin Auth,Refactor Code,Seperate api generator
Improvements
I set up an auth mechanism for admin and used distributed redis and session.

Refactor

I deleted quartz and will switch it completely to hangfire and it will be in a different service.

I moved the generate api to a library, it was easier to use

I separated API and MVC to make it easier to use in the future.

I added a temporary error page

I removed EventSourcing, I will add it in the next commit.